requirements.txt
# Linter
flake8
flake8-bugbear
flake8-builtins
flake8-commas
flake8-import-order
flake8-requirements
pep8-naming
pycodestyle

# Telegram Bot Control
pytelegrambotapi

# .end variables
python-dotenv
